Fix too strict option value parsing
Prior to this fix the below command line input wouldn't validate as
expected:

    $ example --verbose file1.txt file2.txt

The expected behaviour would be that the "--verbose" flag is treated
as boolean true, while "file1.txt" and "file2.txt" would be argument
values. However, the code wrongfully interpreted "file1.txt" as a
value for the "--verbose" option.