Enforce strict definition order validation
Prior to this change options were validated separately from arguments,
and options were always validated first.

After this change the validation is performed in the order of
definition. In practice this means that if the internal state machine
is in such a state that a given value can be interpreted both as an
option flag and as an argument, then the order of definition decides
which of the two will be assigned the value (as opposed to the option
having precedence).