Update internal Repository API
On this level in the architecture the granularity of the term "Define"
doesn't serve any purpose. The term "Save" is deemed to make much more
sense, hence, the "DefineOption" and "DefineArgument" functions are
renamed to "SaveOption" and "SaveArgument".

In solidarity and to maintain consistency, the "SetOptionValue" and
"AddArgumentValue" functions are also renamed to "SaveOptionValue" and
"SaveArgumentValue".